smartcontract/scripts/requirements.txt
web3
eth-account
python-dotenv
pysha3